// src/generated/models/cr2b6_systemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface cr2b6_systems {
  cr2b6_description: string;
  cr2b6_order?: number;
  cr2b6_systemid: string;
  cr2b6_tag?: string;
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: Date;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: number;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: Date;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: Date;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: string;
  statecodename?: string;
  statuscode?: string;
  statuscodename?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  versionnumber?: number;
}